Rename consent video template file
